# Conflicts:
#	360Accounting.Core/IRepository/IGLHeaderRepository.cs
#	360Accounting.Core/IService/IGLHeaderService.cs
#	360Accounting.Data/Repositories/GLHeaderRepository.cs
#	360Accounting.Service/Services/GLHeaderService.cs
#	360Accounting.Web/Common/DataProvider.cs
#	360Accounting.Web/Controllers/JVController.cs

+++ 360Accounting.Core/Entities/JV.cs
@@ -26,11 +26,11 @@ namespace _360Accounting.Core.Entities
         public long Id { get; set; }
         public long HeaderId { get; set; }
         public long CodeCombinationId { get; set; }
-        public double EnteredDr { get; set; }
-        public double EnteredCr { get; set; }
-        public double AccountedDr { get; set; }
-        public double AccountedCr { get; set; }
-        public double Qty { get; set; }
+        public decimal EnteredDr { get; set; }
+        public decimal EnteredCr { get; set; }
+        public decimal AccountedDr { get; set; }
+        public decimal AccountedCr { get; set; }
+        public decimal Qty { get; set; }
         public string Description { get; set; }
         public long TaxRateCode { get; set; }
     }
